Rename datastore model to keyvalue model
