+++ st2actions/tests/unit/test_actionchain.py
@@ -23,9 +23,9 @@ from st2common.constants.action import LIVEACTION_STATUS_SUCCEEDED
 from st2common.constants.action import LIVEACTION_STATUS_FAILED
 from st2common.models.api.notification import NotificationsHelper
 from st2common.models.db.liveaction import LiveActionDB
-from st2common.models.db.datastore import KeyValuePairDB
+from st2common.models.db.keyvalue import KeyValuePairDB
 from st2common.models.system.common import ResourceReference
-from st2common.persistence.datastore import KeyValuePair
+from st2common.persistence.keyvalue import KeyValuePair
 from st2common.persistence.runner import RunnerType
 from st2common.services import action as action_service
 from st2common.util import action_db as action_db_util
