+++ st2reactor/tests/unit/test_data_transform.py
@@ -16,8 +16,8 @@
 import copy
 
 from st2tests import DbTestCase
-from st2common.models.db.datastore import KeyValuePairDB
-from st2common.persistence.datastore import KeyValuePair
+from st2common.models.db.keyvalue import KeyValuePairDB
+from st2common.persistence.keyvalue import KeyValuePair
 from st2reactor.rules import datatransform
