+++ st2tests/st2tests/base.py
@@ -33,7 +33,7 @@ import st2common.models.db.rule as rule_model
 import st2common.models.db.sensor as sensor_model
 import st2common.models.db.trigger as trigger_model
 import st2common.models.db.action as action_model
-import st2common.models.db.datastore as datastore_model
+import st2common.models.db.keyvalue as keyvalue_model
 import st2common.models.db.runner as runner_model
 import st2common.models.db.execution as execution_model
 import st2common.models.db.executionstate as executionstate_model
@@ -59,7 +59,7 @@ ALL_MODELS.extend(rule_model.MODELS)
 ALL_MODELS.extend(sensor_model.MODELS)
 ALL_MODELS.extend(trigger_model.MODELS)
 ALL_MODELS.extend(action_model.MODELS)
-ALL_MODELS.extend(datastore_model.MODELS)
+ALL_MODELS.extend(keyvalue_model.MODELS)
 ALL_MODELS.extend(runner_model.MODELS)
 ALL_MODELS.extend(execution_model.MODELS)
 ALL_MODELS.extend(executionstate_model.MODELS)
